// client/libs/types.ts
import type { ChangeEvent, Dispatch, SetStateAction, MouseEvent } from 'react';

export type AddExpensePayload = {
  onChange: (e: ChangeEvent<HTMLInputElement | HTMLSelectElement>) => void;
  husband_name: string;
  bride_name: string;
  wedding_at: Date;
  event_at: string;
  setStartDate: Dispatch<SetStateAction<Date>>;
  rental_husband: string;
  rental_bride: string;
  sword_husband: string;
  sword_bride: string;
  glove_husband: string;
  glove_bride: string;
  bouquet_husband: string;
  bouquet_bride: string;
  ceremony_husband: string;
  ceremony_bride: string;
  company_husband: string;
  company_bride: string;
  rooftop_husband: string;
  rooftop_bride: string;
  owner_woman_husband: string;
  owner_woman_bride: string;
  owner_man_husband: string;
  owner_man_bride: string;
  select_husband: string;
  select_bride: string;
  frame_husband: string;
  frame_bride: string;
  dress_husband: string;
  dress_bride: string;
  hairpin_husband: string;
  hairpin_bride: string;
  wig_husband: string;
  wig_bride: string;
  video_husband: string;
  video_bride: string;
  etc_husband: string;
  etc_bride: string;
  hanbok_pre_husband: string;
  hanbok_pre_bride: string;
  hanbok_post_husband: string;
  hanbok_post_bride: string;
  play_husband: string;
  play_bride: string;
  anthem_husband: string;
  anthem_bride: string;
  moderator_husband: string;
  moderator_bride: string;
  officiate_husband: string;
  officiate_bride: string;
  meals: string;
  meals_price: string;
  meals_num_husband: string;
  meals_num_bride: string;
  present: string;
  present_price: string;
  present_num_husband: string;
  present_num_bride: string;
  reserve: string;
  reserve_pay: string;
  prepayment_husband: string;
  prepayment_bride: string;
  onBack: () => void;
  onAddExpense: (e: MouseEvent) => void;
};

export type UpdateExpensePayload = {
  id: string;
} & AddExpensePayload;